remove runner request list
